Refactor storage folder structure
