+++ app/routes/cdo/view/certificate.js
@@ -30,7 +30,7 @@ module.exports = [
     options: {
       validate: {
         payload: Joi.object({
-          indexNumber: Joi.string().required(),
+          indexNumber: Joi.string().required()
         }),
         failAction: async (request, h, error) => {
           return h.response().code(400).takeover()
